3rd deliverable/requirements.txt
python-dotenv
pandas_market_calendars
yfinance
ipython
ipykernel
apache-airflow-providers-papermill>=2.0.0
